Add 'list_backends' rather than 'list_devices_between', for more efficient querying.
The old function was really used when backends were always in well-known domains. We want debug commands to be able to show us the state of the system, without knowing very much about it. The new function lists all the backends it can find; it doesn't say anything about whether any are trusted/ valid/ good/ whatever.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>