[xenops] switch to dynamic plugin registration and selection by name
Plugins may be conditionally-compiled (e.g. depending on which libraries
are available) and will register themselves with the Xenops_server module
using a top-level binding.

Note it is critical that the Xenopsd module is linked after all the
plugins, otherwise it will look up a plugin by name and find that none
have yet registered.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>