Refactor the cancellation infrastructure, needed to fix the domain shutdown + cancel race
If a domain shuts down then the xenstore watching thread will attempt to
cancel any outstanding watches. However it is possible for a watch to
be created just after that, which will avoid cancellation.

The fix for this bug will involve watching two paths, so this first commit
hides the details from the clients and imposes a documented interface on top.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>