[xapi] avoid fetching unnecessary events from xenopsd when monitoring a task
If we call UPDATES.get None then we get an event on every object in the system.
Instead we only need to:

  last_id <- UPDATES.last_id
  TASK.stat
  UPDATES.from last_id

which will fetch only new events, and only in the case where the task hasn't
already finished.

With this change a "xe vm-shutdown --force" dropped from ~34s to ~4s in the
presence of 300 VMs, 900 VDIs and 300 VIFs.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>