CA-85940: on error in suspend, clean up the xapi metadata and force-shutdown the domain if it has already 'suspended'
The xenops event thread in xapi now handles the 'power_state = Suspended' case by
pulling the metadata from xenopsd and calling VM.remove.

If a failure happens during suspend, xapi will wait for the event sync
and check the power-state. If the VM is suspended then it will be
force_shutdown because the failure will have left the VM in an invalid
state (probably through a missing or truncated suspend VDI). If the
VM is in any other state then it can be safely left.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>