CA-85940: VIF.unplug now passes the cancellation test
This required a redefinition of what "plugged" means.
We now consider a VIF to be "plugged" if the resources
are still allocated by the host, as reflected by
  /xapi/.../
in xenstore.

When the udev remove event fires and the "hotplug-status"
node is removed, we consider the device as needing an
explicit unplug via the "device action request" mechanism.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>