CA-85940: never directly remove VIF/VBD metadata in failure paths
A background thread constantly synchronises XenAPI metadata with xenopsd
metadata, so it doesn't make sense to remove objects which will be
immediately re-added.

It is ok to 'refresh' the metadata in a {VIF,VBD}.plug operation.

Redefinite what 'plugged' means for VIFs so that a VIF is always
plugged while there is still pending cleanup action (e.g. in xenstore).
This avoids a cancelled unplug leaving stale state lying around.

Clarify the "epoch_{begin,end}" APIs -- these need to know the disk
but not the whole Vbd.t

Signed-off-by: David Scott <dave.scott@eu.citrix.com>